requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
chardet==3.0.4
cryptography==2.6.1
docutils==0.17
entrypoints==0.3
future==0.18.2
idna==2.6
iso8601==0.1.14
keyring==17.1.1
keyrings.alt==3.1.1
Kivy==2.0.0
Kivy-Garden==0.1.4
Pillow==8.2.0
pycrypto==2.6.1
Pygments==2.8.1
pyxdg==0.25
PyYAML==5.4.1
requests==2.21.0
SecretStorage==2.3.1
pyserial==3.5
six==1.12.0
ssh-import-id==5.7
urllib3==1.24.1
pyusb==1.1.1
